requirements.txt
asgiref==3.4.1
Django
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-js-asset==1.2.2
django-safedelete==1.0.0
djangorestframework==3.12.4
pep517==0.10.0
pip-tools==6.2.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6
pandas~=1.3.4
numpy~=1.21.3
